2022/05/05_test.go
package aoc202205

import (
	"testing"

	"github.com/arylatt/advent-of-code/elves"
	"github.com/stretchr/testify/assert"
)

const (
	// Sample1Answer is the expected answer for sample 1
	Sample1Answer = "CMZ"

	// Sample2Answer is the expected answer for sample 2
	Sample2Answer = "MCD"
)

func TestPart1Sample(t *testing.T) {
	td, err := elves.SampleFileToTestData(Sample1Answer)

	if assert.NoError(t, err) {
		elves.TestSample(t, td, Part1)
	}
}

func TestPart1(t *testing.T) {
	elves.TestReal(t, Part1, "2022", "5")
}

func TestPart2Sample(t *testing.T) {
	td, err := elves.SampleFileToTestData(Sample2Answer)

	if assert.NoError(t, err) {
		elves.TestSample(t, td, Part2)
	}
}

func TestPart2(t *testing.T) {
	elves.TestReal(t, Part2, "2022", "5")
}

func TestParseMap(t *testing.T) {
	input := []string{
		"    [D]    ",
		"[N] [C]    ",
		"[Z] [M] [P]",
		" 1   2   3 "}

	expected := Stacks{
		"1": []string{"Z", "N"},
		"2": []string{"M", "C", "D"},
		"3": []string{"P"},
	}

	assert.Equal(t, expected, ParseMap(input))
}
